+++ changelog.md
@@ -28,6 +28,7 @@
 * #625 Bundles can be ordered if an attribute combination of a bundle item is not available
 * Minor UI tweaks in checkout process
 * #666 Export addresses in customer export
+* New shopping cart setting ShowItemsFromWishlistToCartButton
 
 ### Bugfixes
 * Instant search box did not display all results when SQL Fulltext Search was enabled